Added a separated package for router
